+++ src/lmstudio_sdk/__init__.py
@@ -64,6 +64,7 @@ from .utils import (
     BufferedEvent,
     ChannelError,
     lms_default_ports,
+    LiteralOrCoroutine,
     generate_random_base64,
     get_logger,
     RECV,
@@ -105,6 +106,7 @@ __all__ = [
     "KVConfigLayerName",
     "KVConfigStack",
     "KVConfigStackLayer",
+    "LiteralOrCoroutine",
     "LLMApplyPromptTemplateOpts",
     "LLMChatHistory",
     "LLMChatHistoryMessage",
