+++ src/lmstudio_sdk/backend/namespaces/SystemNamespace.py
@@ -3,20 +3,10 @@ from typing import List
 from ...dataclasses import DownloadedModel
 from ...utils import LiteralOrCoroutine
 from ..communications import BaseClientPort
+from .BaseNamespace import BaseNamespace
 
 
-class SystemNamespace:
-    _port: BaseClientPort
-
-    def __init__(self, port: BaseClientPort):
-        self._port = port
-
-    def connect(self) -> None:
-        return self._port.connect()
-
-    def close(self) -> None:
-        return self._port.close()
-
+class SystemNamespace(BaseNamespace[BaseClientPort]):
     def list_downloaded_models(self) -> LiteralOrCoroutine[List[DownloadedModel]]:
         """
         List all the models that have been downloaded.
